+++ CHANGELOG.rst
@@ -9,14 +9,18 @@ Current
 
 New Features
 ^^^^^^^^^^^^
+- Add multi-operator unsup adversarial loss (:gh:`409` by `Andrew Wang`_)
 
 Changed
 ^^^^^^^
+
+- Adversarial losses can be cleanly trained in normal Trainer (:gh:`409` by `Andrew Wang`_)
+- Deprecate adversarial trainer (:gh:`409` by `Andrew Wang`_)
 - (Breaking) Make HDF5Dataset similar to Trainer in the unsupervised setting by using NaNs for ground truths instead of a copy of the measurements (:gh:`761` by `Jérémy Scanvic`_)
 
 Fixed
 ^^^^^
-
+- Fix UAIR loss (:gh:`409` by `Andrew Wang`_)
 
 v0.3.4
 ------